+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### 1.7.20 (build 21046, api 8, 2023-06-04)
+### 1.7.20 (build 21047, api 8, 2023-06-05)
 
 - This seems like a good time for a `refactoring` release in anticipation of
   changes coming in 1.8. Basically this means that a lot of things will be